# GifImagePlugin seek-related optimizations