@AnonymouX47 2022-02-20
For what it's worth, I realized the deficiencies in the course of working with a large (2188 frames, ~8 MiB) GIF image on a project I'm developing.