@radarhere 2022-02-21
Could you upload a copy of your test image? Not to be included in PR, but just as a comment here, so that we can run your speed tests in the same way you are.