@AnonymouX47 2022-02-21
Oh, sorry. Thought it was unnecessary since it should apply to any GIF.
Here's the image:

![TIKTOK_LOGO_400.gif](https://user-images.githubusercontent.com/61663146/154889170-465020ea-0066-4cca-bb10-c12e94a9e87e.gif)