@radarhere 2022-02-21
Hi. So, an animated GIF does not have to be like a... PowerPoint slideshow, where independent images display one after another.
It can be like a series of Photoshop layers. Some layers are mostly transparent, and by applying one layer after another, the image is transformed.

When Pillow seeks back to the first frame each time, it is starting from the bottom of those layers, and applying each new layer in turn. Your solution potentially skips a number of layers, meaning the final image may be incorrect.

Consider this code, adapted from one of your tests.

```python
from PIL import Image

img = Image.open("TIKTOK_LOGO_400.gif")
img.seek(28)  # Mid frame
img.save("before_n_frames.png")
img.n_frames  # Seek to end and reset to frame 28
img.save("after_n_frames.png")
```

before_n_frames.png and after_n_frames should be the same image. Except with your PR, here is what I get.

| before_n_frames.png | after_n_frames.png |
| --- | --- |
| ![before_n_frames](https://user-images.githubusercontent.com/3112309/154895073-39440408-3066-4b0c-a5f6-2545a2d41452.png) | ![after_n_frames](https://user-images.githubusercontent.com/3112309/154895081-e70fb3b3-66c4-46c1-933f-a668fad6e0ff.png) |

With the main branch, they are the same.

So unfortunately, there is a fundamental problem here.