@AnonymouX47 2022-02-21
Indeed, there's a fundamental problem.

Invoking `im.is_animated` or `im.n_frames` while on frame 0 causes any subsequent call to `im.load()` to raise a `ValueError: Unrecognized image mode`.

I guess i didn't realize all these earlier because the changes seemed to work fine with my use case.