@radarhere 2022-02-21
So, my first commit in #6077 helped `test_is_animated_2` - if the user has seeked past the first frame for us, then we don't need to seek to the second frame to know that the image is animated.

I've now added a commit to #6077 that will check if the next byte (or lack thereof) ends the GIF image, and if so, just stay on the current frame. This should eliminate the reset for most cases - I could deliberately create an image that had some data past the last frame and so still needed to be reset, but I would hope that is rare, since I don't know why a GIF writer would do that.

Testing those changes, `test_seek_2`, `test_nframes_1` and `test_nframes_3` now have the same speeds as your PR. `test_seek_EOF_nframes` and `test_nframes_2` are not as fast as in your version, but there are improvements.