@radarhere 2022-02-21
I've pushed another commit - when performing seeks for `n_frames` or `is_animated`, do not actually load the images. We're not interested in them for those moments. This matches this PR's speeds for `test_seek_EOF_nframes` and `test_nframes_2`, and is faster than this PR for `test_nframes_1`.

So somehow, I think that should cover all of the significant differences in speeds from your tests here.