@radarhere 2022-02-21
To be fair, there were 4 failures in our test suite.
But I will grant you, none of them were in test_file_gif.py. So I have created PR #6080